refactor app details
